+++ pkg/util/quotapool/int_rate.go
@@ -32,7 +32,7 @@ type RateLimiter struct {
 // token bucket which has a maximum capacity of burst. If a request attempts to
 // acquire more than burst, it will block until the bucket is full and then
 // put the token bucket in debt.
-func NewRateLimiter(name string, rate Limit, burst int64, options ...Option) *RateLimiter {
+func NewRateLimiter(name string, rate Limit, burst float64, options ...Option) *RateLimiter {
 	rl := &RateLimiter{}
 	bucket := rateBucket{
 		limitConfig: limitConfig{
@@ -50,7 +50,7 @@ func NewRateLimiter(name string, rate Limit, burst int64, options ...Option) *Ra
 
 // Acquire acquires n quota from the RateLimiter. This acquired quota may be
 // released back into the token bucket or it may be consumed.
-func (rl *RateLimiter) Acquire(ctx context.Context, n int64) (*RateAlloc, error) {
+func (rl *RateLimiter) Acquire(ctx context.Context, n float64) (*RateAlloc, error) {
 	if err := rl.WaitN(ctx, n); err != nil {
 		return nil, err
 	}
@@ -59,7 +59,7 @@ func (rl *RateLimiter) Acquire(ctx context.Context, n int64) (*RateAlloc, error)
 
 // WaitN acquires n quota from the RateLimiter. This acquisition cannot be
 // released.
-func (rl *RateLimiter) WaitN(ctx context.Context, n int64) error {
+func (rl *RateLimiter) WaitN(ctx context.Context, n float64) error {
 	if n == 0 {
 		// Special case 0 acquisition.
 		return nil
@@ -75,7 +75,7 @@ func (rl *RateLimiter) WaitN(ctx context.Context, n int64) error {
 // AdmitN acquire n quota from the RateLimiter if it succeeds. It will return
 // false and not block if there is currently insufficient quota or the pool is
 // closed.
-func (rl *RateLimiter) AdmitN(n int64) bool {
+func (rl *RateLimiter) AdmitN(n float64) bool {
 	r := rl.newRateRequest(n)
 	defer rl.putRateRequest(r)
 	return rl.qp.Acquire(context.Background(), (*rateRequestNoWait)(r)) == nil
@@ -87,7 +87,7 @@ func (rl *RateLimiter) AdmitN(n int64) bool {
 // update to 20, the quota will increase to 15. Similarly, if the burst is
 // decreased by 10, the current quota will decrease accordingly, potentially
 // putting the limiter into debt.
-func (rl *RateLimiter) UpdateLimit(rate Limit, burst int64) {
+func (rl *RateLimiter) UpdateLimit(rate Limit, burst float64) {
 	rl.qp.Update(func(res Resource) (shouldNotify bool) {
 		r := res.(*rateBucket)
 		shouldNotify = r.burst < burst || r.rate < rate
@@ -110,13 +110,13 @@ type rateBucket struct {
 
 type limitConfig struct {
 	rate  Limit
-	burst int64
+	burst float64
 }
 
 // RateAlloc is an allocated quantity of quota which can be released back into
 // the token-bucket RateLimiter.
 type RateAlloc struct {
-	alloc int64
+	alloc float64
 	rl    *RateLimiter
 }
 
@@ -144,7 +144,7 @@ func (ra *RateAlloc) Consume() {
 type rateAlloc RateAlloc
 
 type rateRequest struct {
-	want int64
+	want float64
 }
 
 var rateRequestSyncPool = sync.Pool{
@@ -153,7 +153,7 @@ var rateRequestSyncPool = sync.Pool{
 
 // newRateRequest allocates a rateRequest from the sync.Pool.
 // It should be returned with putRateRequest.
-func (rl *RateLimiter) newRateRequest(v int64) *rateRequest {
+func (rl *RateLimiter) newRateRequest(v float64) *rateRequest {
 	r := rateRequestSyncPool.Get().(*rateRequest)
 	*r = rateRequest{want: v}
 	return r
@@ -164,6 +164,7 @@ func (rl *RateLimiter) putRateRequest(r *rateRequest) {
 	rateRequestSyncPool.Put(r)
 }
 
+// Acquire is part of the Request interface.
 func (i *rateRequest) Acquire(
 	ctx context.Context, res Resource,
 ) (fulfilled bool, tryAgainAfter time.Duration) {
@@ -206,6 +207,7 @@ func (r *rateBucket) update(now time.Time) {
 	}
 }
 
+// ShouldWait is part of the Request interface.
 func (i *rateRequest) ShouldWait() bool {
 	return true
 }
@@ -214,7 +216,7 @@ var rateAllocSyncPool = sync.Pool{
 	New: func() interface{} { return new(rateAlloc) },
 }
 
-func (rl *RateLimiter) newRateAlloc(v int64) *rateAlloc {
+func (rl *RateLimiter) newRateAlloc(v float64) *rateAlloc {
 	a := rateAllocSyncPool.Get().(*rateAlloc)
 	*a = rateAlloc{alloc: v, rl: rl}
 	return a
@@ -229,12 +231,14 @@ func (rl *RateLimiter) putRateAlloc(a *rateAlloc) {
 // quota.
 type rateRequestNoWait rateRequest
 
+// Acquire is part of the Request interface.
 func (r *rateRequestNoWait) Acquire(
 	ctx context.Context, resource Resource,
 ) (fulfilled bool, tryAgainAfter time.Duration) {
 	return (*rateRequest)(r).Acquire(ctx, resource)
 }
 
+// ShouldWait is part of the Request interface.
 func (r *rateRequestNoWait) ShouldWait() bool {
 	return false
 }
